memory_cycle/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
plotly>=5.18.0
pyyaml>=6.0
